requirements.txt
cachetools==2.0.1
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
google-api-core==1.1.2
google-api-python-client==1.6.7
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.1.0
google-cloud-core==0.28.1
google-cloud-storage==1.8.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
httplib2==0.11.3
idna==2.6
nose==1.3.7
oauth2client==4.1.2
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybuilder==0.11.17
pyminizip==0.2.3
pypandoc==1.3.3
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
tailer==0.4.1
tblib==1.3.2
unittest-xml-reporting==2.2.0
uritemplate==3.0.0
urllib3==1.22
pycryptodome==3.6.4
